chore: use literal union type instead of enums
In order to be able to generate two versions of the codegen:
- one for the backend generated from the "admin" permissions
- one for the frontend generated from the "org_owner" permissions

we need to replace enums by literals unions so that we can share some code between the 2 SDKs. Indeed, when using enums, Typescript complains about not having any overlap between enums even if they have the same values:
<img width="988" alt="image" src="https://user-images.githubusercontent.com/2678610/161303594-c9030409-a855-487e-94a8-dd6448a7a2c9.png"> ([code](https://www.typescriptlang.org/play?#code/KYDwDg9gTgLgBMAdgVwLYJag+gZwJapgA2wcA3gLABQccAxgIY7ACMcAvHAOSPMtcAaarV7AATB26ixg6gF9qoSLAxpV2fIRJYJlGvSatJPQ-yH7px6bKoKq1OhEQ4IJAHREIAcwAUSNLgExMBuomzsEeqBWsA6AJRAA))

While string literals unions can simply be compared:
<img width="290" alt="image" src="https://user-images.githubusercontent.com/2678610/161303751-b0d9bc5f-7b7b-46db-b926-553a9faba1ad.png">

Another benefit of using string literals unions is that they are non generated to the build output.